test(memory): split importer resolution tests into tests/claude_mem/
The repeated-separator regression tests pushed
tests/test_claude_mem_scripts.py to 507 lines, past the 500-line taste-lint
ceiling. That is not advisory at the gate: taste-count-ratchet reported
576 > baseline 575 and pre_pr.py failed, so the file-size finding blocks a
push rather than merely warning.

I flagged this exact split as needed last round and then made the addition
without doing it, so the ceiling caught what I had already predicted.

The importer tests move to tests/claude_mem/, the package that already holds
claude-mem test modules, rather than a sibling file next to the export tests.
Both modules now sit under the ceiling (129 and 413 lines) and the split falls
on a real seam: export-script behavior on one side, importer path resolution
and exit codes on the other.

The loader helper is duplicated rather than hoisted to conftest. It is ten
lines, and the two copies differ (the new one walks up one more directory and
loads a single module), so extracting it would need a parameterized shared
helper for no reader benefit.

Refs #4780
